+++ .test-e2e/xp-image/config/com.enonic.app.contentstudio.cfg
