+++ .test-e2e/xp-image/config/com.enonic.xp.content.cfg
