+++ .test-e2e/xp-image/config/no.nav.navno.cfg
