+++ .test-e2e/xp-image/setenv.sh
