+++ .test-e2e/xp-test-data-app/.babelrc.js
@@ -1,0 +1,3 @@
+export default {
+    presets: ['@babel/preset-env', '@babel/preset-typescript'],
+};
