+++ .test-e2e/xp-test-data-app/gradle.properties
@@ -1,0 +1,10 @@
+#
+# Project settings
+#
+group = no.nav
+projectName = navno-testdata
+appName = no.nav.navno.testdata
+# xpVersion set in deploy workflows overrides this value
+xpVersion = 7.14.4
+applicationKey = no.nav.navno.testdata
+version = 1.0.0-LOCAL
