+++ .test-e2e/xp-test-data-app/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = projectName
