+++ .test-e2e/xp-test-data-app/src/main/resources/application.xml
@@ -1,0 +1,3 @@
+<application>
+  <description>nav.no test data</description>
+</application>
