+++ build.gradle
@@ -1,11 +1,11 @@
 plugins {
-    id 'java'
     id 'com.enonic.xp.app' version '3.4.0'
     id "com.github.node-gradle.node" version '7.0.0'
     id 'no.item.xp.codegen' version '2.3.0'
 }
 
 repositories {
+    mavenLocal()
     mavenCentral()
     xp.enonicRepo()
 }
