+++ src/main/resources/lib/contentUpdate/content-data-locale-fallback-update.ts
@@ -6,7 +6,7 @@ import { ArrayOrSingle } from '../../types/util-types';
 import { ContentDataLocaleFallback } from '@xp-types/site/content-types/content-data-locale-fallback';
 import { runInContext } from '../context/run-in-context';
 import { logger } from '../utils/logging';
-import { SUPER_USER_FULL } from '../constants';
+import { SUPER_USER_PRINCIPAL } from '../constants';
 
 type FallbackContent = Content<'no.nav.navno:content-data-locale-fallback'>;
 type Item = NonNullable<ContentDataLocaleFallback['items']>[number];
@@ -122,7 +122,7 @@ export const contentDataLocaleFallbackRefreshItems = (content: FallbackContent) 
     }
 
     // Check the last modifier to ensure this function never runs in an infinite loop
-    if (content.modifier === SUPER_USER_FULL) {
+    if (content.modifier === SUPER_USER_PRINCIPAL) {
         logger.warning(
             `Possible update loop on update handler for locale fallback data: ${content._id}`
         );
