+++ tsconfig.json
@@ -44,5 +44,5 @@
             "/assets/dayjs/1.11.9/plugin/utc.js": ["../../../node_modules/dayjs/plugin/utc"]
         }
     },
-    "include": ["src/main/resources/**/*.ts", ".xp-codegen/**/*.ts"]
+    "include": ["src/main/resources/**/*.ts", ".xp-codegen/**/*.ts", "global.d.ts"]
 }
